fix(Disasters panel): 20467 Show feed selector and bbox filter in Empty state of Disasters panel
