+++ src/features/events_list/components/EventsPanelErrorMessage/EventsPanelErrorMessage.module.css
@@ -1,0 +1,3 @@
+.container {
+  width: 100%;
+}
